update target android and make score handlers work with longs

+++ cocos2dx/platform/android/CCGameServices.cpp
@@ -42,7 +42,7 @@ void CCGameServices::showLeaderboard() {
     //CCAssert(false, "implement CCGameServices::showLeaderboard");
 }
 
-void CCGameServices::submitScore(const char *category, long long score) {
+void CCGameServices::submitScore(const char *category, long score) {
     CCLog("This needs implemented");
     submitScoreJNI(category, score);
     CCAssert(false, "implement CCGameServices::submitScore");
